기존의 예외 처리 클래스들을 삭제하고, 새로운 패키지 구조에 맞춰 재구성된 예외 처리 클래스들을 추가함
