--- src/main/java/com/luckeat/luckeatbackend/common/exception/BadRequestException.java
@@ -1,11 +1,0 @@
-package com.luckeat.luckeatbackend.common.exception;
-
-public class BadRequestException extends CustomException {
-	public BadRequestException(String message) {
-		super(ErrorCode.BAD_REQUEST, message);
-	}
-
-	public BadRequestException() {
-		super(ErrorCode.BAD_REQUEST);
-	}
-}
