--- src/main/java/com/luckeat/luckeatbackend/common/exception/CustomException.java
@@ -1,18 +1,0 @@
-package com.luckeat.luckeatbackend.common.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-	private final ErrorCode errorCode;
-
-	public CustomException(ErrorCode errorCode) {
-		super(errorCode.getMessage());
-		this.errorCode = errorCode;
-	}
-
-	public CustomException(ErrorCode errorCode, String message) {
-		super(message);
-		this.errorCode = errorCode;
-	}
-}
