--- src/main/java/com/luckeat/luckeatbackend/common/exception/ForbiddenException.java
@@ -1,11 +1,0 @@
-package com.luckeat.luckeatbackend.common.exception;
-
-public class ForbiddenException extends CustomException {
-	public ForbiddenException(String message) {
-		super(ErrorCode.FORBIDDEN, message);
-	}
-
-	public ForbiddenException() {
-		super(ErrorCode.FORBIDDEN);
-	}
-}
