+++ src/main/java/com/luckeat/luckeatbackend/common/exception/GlobalExceptionHandler.java
@@ -13,6 +13,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
+import com.luckeat.luckeatbackend.common.exception.base.CustomException;
+
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
