--- src/main/java/com/luckeat/luckeatbackend/common/exception/NotFoundException.java
@@ -1,11 +1,0 @@
-package com.luckeat.luckeatbackend.common.exception;
-
-public class NotFoundException extends CustomException {
-	public NotFoundException(String message) {
-		super(ErrorCode.NOT_FOUND, message);
-	}
-
-	public NotFoundException() {
-		super(ErrorCode.NOT_FOUND);
-	}
-}
