+++ src/main/java/com/luckeat/luckeatbackend/common/exception/product/ProductForbiddenException.java
@@ -1,6 +1,6 @@
 package com.luckeat.luckeatbackend.common.exception.product;
 
-import com.luckeat.luckeatbackend.common.exception.ForbiddenException;
+import com.luckeat.luckeatbackend.common.exception.base.ForbiddenException;
 
 public class ProductForbiddenException extends ForbiddenException {
 	public ProductForbiddenException() {
