+++ src/main/java/com/luckeat/luckeatbackend/common/exception/product/ProductNameDuplicateException.java
@@ -1,6 +1,6 @@
 package com.luckeat.luckeatbackend.common.exception.product;
 
-import com.luckeat.luckeatbackend.common.exception.ConflictException;
+import com.luckeat.luckeatbackend.common.exception.base.ConflictException;
 
 public class ProductNameDuplicateException extends ConflictException {
 	public ProductNameDuplicateException() {
