+++ src/main/java/com/luckeat/luckeatbackend/common/exception/review/ReviewNotFoundException.java
@@ -1,6 +1,6 @@
 package com.luckeat.luckeatbackend.common.exception.review;
 
-import com.luckeat.luckeatbackend.common.exception.NotFoundException;
+import com.luckeat.luckeatbackend.common.exception.base.NotFoundException;
 
 public class ReviewNotFoundException extends NotFoundException {
 	public ReviewNotFoundException() {
