+++ src/main/java/com/luckeat/luckeatbackend/common/exception/user/NicknameDuplicateException.java
@@ -1,6 +1,6 @@
 package com.luckeat.luckeatbackend.common.exception.user;
 
-import com.luckeat.luckeatbackend.common.exception.ConflictException;
+import com.luckeat.luckeatbackend.common.exception.base.ConflictException;
 
 public class NicknameDuplicateException extends ConflictException {
 	public NicknameDuplicateException() {
